### 2017-11-24 #160 @steffilazerte
weathercan
`weathercan` helps users access historical weather data from Environment and Climate Change Canada (ECCC). Data is freely available from the ECCC website, but downloading and processing the data by hand can be time consuming. `weathercan` automatically downloads and formats data from multiple stations and over large date ranges.

```
Package: weathercan
Title: Download Weather Data from the Environment and Climate Change Canada Website
Version: 0.2.2.9000
     person("Steffi", "LaZerte", email = "steffi@steffi.ca", role = c("aut","cre")),
     person("Sam", "Albers", email = "sam.albers@gmail.com", role = c("ctb")))
Description: Provides means for downloading historical weather data from
    the Environment and Climate Change Canada website
    (<http://climate.weather.gc.ca/historical_data/search_historic_data_e.html>).
    Data can be downloaded from multiple stations and over large date ranges
    and automatically processed into a single dataset. Tools are also provided
    to identify stations either by name or proximity to a location.
    lubridate (>= 1.7.1),
    magrittr (>= 1.5),
    methods (>= 3.2.2),
    sp (>= 1.2-3),
    stringi (>= 1.1.2),
    tibble (>= 1.2),
    tidyr (>= 0.4.1),
    xml2 (>= 0.1.2)
    devtools,
```

<https://github.com/steffilazerte/weathercan>

**data retrieval**:
`weathercan` downloads multiple data sets at a time from the Environment and Climate Change Canada website.

**data munging**:
`weathercan` tidies and combines the downloaded data into a format ready for analysis in R. `weathercan` also provides tools for conducting linear interpolation for adding weather data to other data sets which might not perfectly align with the timing of weather observations.

Any one who needs access to Canadian weather data: Biologists, Meteorologists, Geographers, Hydrologists, Climatologists, Students, Teachers, etc.

`rclimateca` is a package available on CRAN that is similar to `weathercan`.

Both packages concern the downloading of Environment and Climate Change Canada weather data into an easy format for use in R.

The largest difference between the two packages is that `weathercan` includes functions for interpolating weather data and directly integrating it into other data sources. Another difference is that `weathercan` actively seeks to apply tidy data principles in R and integrates well with the tidyverse including using tibbles and nested listcols. The `weathercan` package also includes a website and tutorials/vignettes. We would also argue that `weathercan` is slightly more user-friendly.

Status: OK

R CMD check results
0 errors | 0 warnings | 0 notes

Best:

  - https://github.com/adamhsparks (similar packages)
  - https://github.com/ivanhanigan (similar work)
  - https://github.com/cvitolo (likely a user of this package or packages like this)

Alternative:

  - https://github.com/bpbond (could be potential user for this package or packages like this)
  - https://github.com/aappling-usgs (could be potential user for this package or packages like this)

- Extra: goodpractice::gp() output:

── GP weathercan ─────────────────────────────────────────────

    R/interpolate.R:64:NA
    R/interpolate.R:65:NA
    R/interpolate.R:84:NA
    R/interpolate.R:142:NA
    R/interpolate.R:151:NA

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80
    characters wide. Try make your lines shorter than 80 characters

    R/interpolate.R:62:1
    R/interpolate.R:63:1
    R/interpolate.R:78:1
    R/interpolate.R:97:1
    R/interpolate.R:119:1

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the input data. Consider using

    R/interpolate.R:103:18
    R/interpolate.R:165:18
    R/interpolate.R:182:18
    R/weather.R:367:10
    R/weather.R:369:30

  ✖ fix this R CMD check NOTE: Note: found 26 marked UTF-8 strings

__Comment:__

This note doesn't pop up when I run `devtools::check()` on my linux machine, so I'm not really sure what's going on. I wouldn't mind some assistance in determining:

a) whether it matters
b) if it does, how to fix it in a cross platform manner
