### 2017-11-24 #160 @sckott
👋 @steffilazerte - thanks for your submission!  We'll discuss fit for rOpenSci and get back to you soon.
