### 2017-11-28 #160 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @steffilazerte  !!

Goodpractice output for you and reviewers to consider:

```r
It is good practice to

    R/interpolate.R:64:NA
    R/interpolate.R:65:NA
    R/interpolate.R:84:NA
    R/interpolate.R:142:NA
    R/interpolate.R:151:NA
    ... and 27 more lines

    R/interpolate.R:62:1
    R/interpolate.R:63:1
    R/interpolate.R:78:1
    R/interpolate.R:97:1
    R/interpolate.R:119:1
    ... and 93 more lines

    vapply() instead.

    R/interpolate.R:103:18
    R/interpolate.R:165:18
    R/interpolate.R:182:18
    R/weather.R:367:10
    R/weather.R:369:30
    ... and 4 more lines

  ✖ fix this R CMD check NOTE: Note: found 26 marked UTF-8
    strings
```

Seeking reviewers now.

---

Reviewers:  xx xx
Due date: xxxx
