### 2017-11-29 #160 @sckott
Reviewers assigned. They are: @joethorley @softloud
Due date is 2017-12-20
