### 2017-11-29 #160 @joethorley
- [x] As the reviewer I confirm that there are no conflicts of interest for me to review this work.

>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

This is a useful package to locate Environment Canada weather stations,
and download and interpolate the data.

#### A statement of need

The README definitely requires a brief discussion of how the current package differs from `rclimateca` and `CHCN`.

#### Examples

An example is lacking for `stations_all()` - this is minor but technically in violation of the [documentation policy](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md#-documentation).

The demonstration of the use of `dplyr::select` is unnecessary in the README.

#### Packaging Guidelines

The function naming should follow the suggested `object_verb()` [naming scheme](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md#funvar) ie `weather()` should become something like `weather_download()`, `stations_all()` to `stations_download()`, and  `add_weather()` to `weather_interpolate()` etc.

The README lacks citation information.

The two most recent NEWS.md headings lack dates.

#### DESCRIPTION

The DESCRIPTION file should include the Date of the current version.

#### Coding Style

The readability would be improved by breaking the bigger functions into subfunctions but this could be alot of work and is not necessary for acceptance of the package.
Only use return() for early returns.
