### 2017-11-29 #160 @steffilazerte
Thanks @joethorley, this all looks pretty straightforward, thanks for your speedy review! I'll start looking into these changes.
