### 2017-12-05 1645a34
Add stations_all example, update documentation
