+++ man/stations_all.Rd
@@ -24,13 +24,24 @@ A tibble containing station names, station ID codes and dates of
 }
 \description{
 This function can be used to download a Station Inventory CSV file from
-Environment Canada. Note that the 'stations' data set included in this
-package contains station data downloaded when the package was last compiled,
-so it may not be necessary to call this function (and this function may take
-a few minutes to run).
+Environment and Climate Change Canada. This is only necessary if the station
+you're interested was only recently added. The 'stations' data set included
+in this package contains station data downloaded when the package was last
+compiled. This function may take a few minutes to run.
 }
 \details{
 URL defaults to
 ftp://client_climate@ftp.tor.ec.gc.ca/Pub/Get_More_Data_Plus_de_donnees/
 Station%20Inventory%20EN.csv unless otherwise specified
+}
+\examples{
+
+\dontrun{
+ # Update stations data frame
+ s <- stations_all()
+
+ # Use new data frame to search for stations
+ stations_search("Winnipeg", stn = s)
+}
+
 }
