### 2017-12-05 f461d07
Remove select example
