### 2017-12-05 9455b01
Remove select example
