+++ DESCRIPTION
@@ -1,6 +1,7 @@
 Package: weathercan
 Type: Package
 Title: Download Weather Data from the Environment and Climate Change Canada Website
+Date: 2017-12-05
 Version: 0.2.2.9000
 Authors@R: c(
      person("Steffi", "LaZerte", email = "steffi@steffi.ca", role = c("aut","cre")),
