### 2017-12-05 84c640b
Add citation info
