### 2017-12-05 a9e2246
Add citation info
