### 2017-12-05 0d7dfd7
Add missing dates to NEWS
