+++ NEWS.md
@@ -4,7 +4,7 @@ author: "Steffi LaZerte"
 output: html_document
 ---
 
-# weathercan 0.2.3
+# weathercan 0.2.3 (2017-11-22)
 
 ## Changes
 - Updated `stations` data
@@ -19,7 +19,7 @@ output: html_document
 - `weather(format = FALSE)` properly returns data
 - updated `weather()` to work with `lubridate` 1.7.1
 
-# weathercan 0.2.2
+# weathercan 0.2.2 (2017-06-16)
 
 ## Changes
 - Update and expand vignettes (closes #15)
