### 2017-12-05 839cccc
Remove non-early explicit returns
