### 2017-12-05 ccbabcc
Remove non-early explicit returns
