### 2017-12-07 e2cbded
remove stations_search() ability to search by coordinate. Removes sp dependency and would close #45
