### 2017-12-07 5a3b6b6
add ropensci badge
