### 2017-12-09 2032561
set eval=FALSE in docs/vignettes that search by coordinates
