+++ README.md
@@ -84,20 +84,19 @@ You can also search by proximity:
 stations_search(coords = c(50.667492, -120.329049), dist = 20, interval = "hour")
 ```
 
-    ## # A tibble: 3 x 13
-    ##     prov station_name station_id climate_id WMO_id  TC_id   lat     lon  elev interval start   end
-    ##   <fctr>        <chr>     <fctr>     <fctr> <fctr> <fctr> <dbl>   <dbl> <dbl>    <chr> <int> <int>
-    ## 1     BC   KAMLOOPS A       1275    1163780  71887    YKA  50.7 -120.44 345.3     hour  1953  2013
-    ## 2     BC KAMLOOPS AUT      42203    1163842  71741    ZKA  50.7 -120.44 345.0     hour  2006  2017
-    ## 3     BC   KAMLOOPS A      51423    1163781  71887    YKA  50.7 -120.45 345.3     hour  2013  2017
-    ## # ... with 1 more variables: distance <dbl>
-
 ### Weather
 
 Once you have your `station_id`(s) you can download weather data:
 
 ``` r
 kam <- weather(station_ids = 51423, start = "2016-01-01", end = "2016-02-15")
+```
+
+    ## Warning: All formats failed to parse. No formats found.
+
+    ## Warning: All formats failed to parse. No formats found.
+
+``` r
 kam
 ```
 
@@ -120,6 +119,17 @@ You can also download data from multiple stations at once:
 
 ``` r
 kam_pg <- weather(station_ids = c(48248, 51423), start = "2016-01-01", end = "2016-02-15")
+```
+
+    ## Warning: All formats failed to parse. No formats found.
+
+    ## Warning: All formats failed to parse. No formats found.
+
+    ## Warning: All formats failed to parse. No formats found.
+
+    ## Warning: All formats failed to parse. No formats found.
+
+``` r
 kam_pg
 ```
 
@@ -157,11 +167,16 @@ Citation
 citation("weathercan")
 ```
 
+    ## Warning in citation("weathercan"): no date field in DESCRIPTION file of package 'weathercan'
+
+    ## Warning in citation("weathercan"): could not determine year for 'weathercan' from package
+    ## DESCRIPTION file
+
     ## 
     ## To cite package 'weathercan' in publications use:
     ## 
-    ##   Steffi LaZerte (2017). weathercan: Download Weather Data from the Environment and
-    ##   Climate Change Canada Website. R package version 0.2.2.9000.
+    ##   Steffi LaZerte (NA). weathercan: Download Weather Data from the Environment and Climate
+    ##   Change Canada Website. R package version 0.2.3.
     ##   https://github.com/steffilazerte/weathercan
     ## 
     ## A BibTeX entry for LaTeX users is
@@ -169,8 +184,7 @@ citation("weathercan")
     ##   @Manual{,
     ##     title = {weathercan: Download Weather Data from the Environment and Climate Change Canada Website},
     ##     author = {Steffi LaZerte},
-    ##     year = {2017},
-    ##     note = {R package version 0.2.2.9000},
+    ##     note = {R package version 0.2.3},
     ##     url = {https://github.com/steffilazerte/weathercan},
     ##   }
 
