### 2017-12-09 0eda57d
remove eval=FALSE; install on ci libs for spatial
