+++ .travis.yml
@@ -6,6 +6,11 @@ cache: packages
 git:
   depth: 150
 
+before_install:
+  - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
+  - sudo apt-get --yes --force-yes update -qq
+  - sudo apt-get install --yes libudunits2-dev libproj-dev libgeos-dev libgdal-dev
+
 matrix:
   include:
     - os: linux
