### 2017-12-09 012557a
revise travis. no osx
