### 2017-12-09 54b8e33
catch warnings in documentation
