+++ README.Rmd
@@ -62,14 +62,14 @@ stations_search(coords = c(50.667492, -120.329049), dist = 20, interval = "hour"
 
 Once you have your `station_id`(s) you can download weather data:
 
-```{r, R.options = list(tibble.max_extra_cols = 0)}
+```{r, warning=FALSE, R.options = list(tibble.max_extra_cols = 0)}
 kam <- weather(station_ids = 51423, start = "2016-01-01", end = "2016-02-15")
 kam
 ```
 
 You can also download data from multiple stations at once:
 
-```{r, R.options = list(tibble.max_extra_cols = 0)}
+```{r, warning=FALSE, R.options = list(tibble.max_extra_cols = 0)}
 kam_pg <- weather(station_ids = c(48248, 51423), start = "2016-01-01", end = "2016-02-15")
 kam_pg
 ```
