### 2017-12-15 #160 @softloud
Thanks for the prompt. I am impressed by the efficiency of rOpenSci! Hoping to get it done this weekend.
