### 2017-12-19 #160 @softloud
# Package Review

* [ ]**Vignette(s)** demonstrating major functionality that runs successfully locally
* [x]**Function Documentation:** for all exported functions in R help
* [ ]**Examples** for all exported functions in R Help that run successfully locally
* [x]**Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

>##### For packages co-submitting to JOSS
>
>* [x]The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>* [x]**A short summary** describing the high-level functionality of the software
>* [x]**Authors:**  A list of authors with their affiliations
>* [x]**A statement of need** clearly stating problems the software is designed to solve and its target audience.
>* [x]**References:** with DOIs for all those that have one (e.g. papers, datasets, software).

* [x]**Installation:** Installation succeeds as documented.
* [x]**Functionality:** Any functional claims of the software been confirmed.
* [x]**Performance:** Any performance claims of the software been confirmed.
* [x]**Automated tests:** Unit tests cover essential functions of the package
* [ ]**Packaging guidelines**: The package conforms to the [ROpenSci packaging guide](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md)

* [ ]**The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**

---

I appreciate the opportunity to review a package like this; I learnt a lot from its well-designed structure. I like the idea of a  `tidy`-friendly `glossary` dataframe with accompanying vignette.

Now I don't know if this is something to do with my inexperience, but I couldn't get the vignettes to run. In fact, I struggled to access them at all.

Just in case this is an OS thing, I'm trying this out on a Macbook Air High Sierra i5 from 2014. Let me know if there's any other information you need from me.
```
> library(weathercan)
> vignette("usage") # No output other than warning.
vignette ‘usage’ not found
> vignette("broom") # The broom vignette opens in the Help pane.
```

#### Examples

As @joethorley noted, there are one or two functions that are missing an example. When I went back through I think it was just `stations_all` that needs a brief example for the function documentation to conform to the rOpenSci package guide.

There was one failure when running tests. This looks very minor, if it is anything to worry about. The tests also appear comprehensive.

```
==> devtools::test()

Loading weathercan
Loading required package: testthat
Testing weathercan
✔ | OK F W S | Context
✖ | 87 1     | Raw interpolation (time) [1.0 s]
───────────────────────────────────────────────────────────────────────
test_interpolate.R:78: failure: approx_na_rm (time) replaces gaps with NAs
sum(is.na(a$y)) not equal to 195.
1/1 mismatches
[1] 0 - 195 == -195
───────────────────────────────────────────────────────────────────────
```

### Other tests

I ran various other ad hoc tests, including the `goodpractice` package, but didn't find anything other than the one error thrown by `testthat` to report. So, barring a few minor tweaks, perhaps, the `weathercan` package seems ready for submission to me.
