### 2017-12-19 #160 @joethorley
FWIW I built the vignettes successfully on OSX High Sierra using `install.packages("~/Code/weathercan/weathercan_0.2.2.9000.tar.gz", repos = NULL, type = "source")`
