### 2017-12-20 #160 @steffilazerte
Great thanks to both of you. I'm not sure what's going on with that test but I'll take a look.
In general, I've been incorporating changes based on your suggestions to the developmental branch and will push to the master branch sometime after Christmas!
