### 2018-01-02 40cb49b
Move sp to suggests and warn user to install for coordinate searches
