+++ DESCRIPTION
@@ -24,7 +24,6 @@ Imports:
     lubridate (>= 1.7.1),
     magrittr (>= 1.5),
     methods (>= 3.2.2),
-    sp (>= 1.2-3),
     stringi (>= 1.1.2),
     tibble (>= 1.2),
     tidyr (>= 0.4.1),
@@ -35,6 +34,7 @@ Suggests:
     ggplot2,
     knitr,
     rmarkdown,
+    sp,
     stringr,
     testthat
 VignetteBuilder: knitr
