### 2018-01-02 d0445e7
Add guides for contributions
