### 2018-01-02 648177d
Add section describing similar packages and link to contributions guide.
