### 2018-01-03 8493e35
Rename functions: - weather_dl (from weather) - stations_dl (from stations_all) - weather_interp (from add_weather) - tz_calc (from get_tz) And update pkgdown site
