### 2018-01-03 6439118
Add information regarding vignettes
