+++ README.md
@@ -20,6 +20,16 @@ install.packages("devtools") # If not already installed
 devtools::install_github("steffilazerte/weathercan") 
 ```
 
+To build the vignettes (tutorials) locally, use:
+
+``` r
+devtools::install_github("steffilazerte/weathercan", build_vignettes = TRUE) 
+```
+
+View the available vignettes with `vignette(package = "weathercan")`
+
+View a particular vignette with, for example, `vignette("weathercan", package = "weathercan")`
+
 General usage
 -------------
 
@@ -34,15 +44,14 @@ head(stations)
 ```
 
     ## # A tibble: 6 x 12
-    ##     prov           station_name station_id climate_id WMO_id  TC_id   lat     lon  elev interval
-    ##   <fctr>                  <chr>     <fctr>     <fctr> <fctr> <fctr> <dbl>   <dbl> <dbl>    <chr>
-    ## 1     BC            ACTIVE PASS         14    1010066   <NA>   <NA> 48.87 -123.28   4.0     hour
-    ## 2     BC            ALBERT HEAD         15    1010235   <NA>   <NA> 48.40 -123.48  17.0     hour
-    ## 3     BC BAMBERTON OCEAN CEMENT         16    1010595   <NA>   <NA> 48.58 -123.52  85.3     hour
-    ## 4     BC             BEAR CREEK         17    1010720   <NA>   <NA> 48.50 -124.00 350.5     hour
-    ## 5     BC            BEAVER LAKE         18    1010774   <NA>   <NA> 48.50 -123.35  61.0     hour
-    ## 6     BC             BECHER BAY         19    1010780   <NA>   <NA> 48.33 -123.63  12.2     hour
-    ## # ... with 2 more variables: start <int>, end <int>
+    ##   prov   station_name           station_id clima… WMO_id TC_id   lat   lon   elev inte… start   end
+    ##   <fctr> <chr>                  <fctr>     <fctr> <fctr> <fct> <dbl> <dbl>  <dbl> <chr> <int> <int>
+    ## 1 BC     ACTIVE PASS            14         10100… <NA>   <NA>   48.9  -123   4.00 hour     NA    NA
+    ## 2 BC     ALBERT HEAD            15         10102… <NA>   <NA>   48.4  -123  17.0  hour     NA    NA
+    ## 3 BC     BAMBERTON OCEAN CEMENT 16         10105… <NA>   <NA>   48.6  -124  85.3  hour     NA    NA
+    ## 4 BC     BEAR CREEK             17         10107… <NA>   <NA>   48.5  -124 350    hour     NA    NA
+    ## 5 BC     BEAVER LAKE            18         10107… <NA>   <NA>   48.5  -123  61.0  hour     NA    NA
+    ## 6 BC     BECHER BAY             19         10107… <NA>   <NA>   48.3  -124  12.2  hour     NA    NA
 
 ``` r
 glimpse(stations)
@@ -70,11 +79,11 @@ stations_search("Kamloops", interval = "hour")
 ```
 
     ## # A tibble: 3 x 12
-    ##     prov station_name station_id climate_id WMO_id  TC_id   lat     lon  elev interval start   end
-    ##   <fctr>        <chr>     <fctr>     <fctr> <fctr> <fctr> <dbl>   <dbl> <dbl>    <chr> <int> <int>
[…]
