### 2018-01-03 aa5a050
Delete ropensci_response.md
