### 2018-01-03 a47800f
Fix link to CODE_OF_CONDUCT.md
