+++ README.md
@@ -206,4 +206,4 @@ We welcome any and all contributions! To make the process as painless as possibl
 Code of Conduct
 ---------------
 
-Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
+Please note that this project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By participating in this project you agree to abide by its terms.
