### 2018-01-04 #160 @joethorley
@sckott I am happy with the authors response - I have no further comments
