### 2018-01-24 #160 @sckott
thanks for reminding me @steffilazerte

was waiting on the other reviewes response. but we'll pass on that.

I'm taking a quick look to see if there's anything else
