### 2018-01-24 #160 @sckott
Just a few things I noticed:

- remove the Date from the DESCRIPTION file - i realize a rewiewer suggested that, but it should be removed - it gets added to the pkg when it's built on the CRAN servers
- goodpractice highlights that you have a lot of lines of code that are longer than 80 characters. that's not a make or break thing, but I'd strongly recommend shortenin lines to 80 width
- gp also highlights use of `sapply()`. Strongly recommend avoiding `sapply`

of the above, can you take care of the 1st and 3rd before we move on?
