### 2018-01-26 560d87d
Remove date
