### 2018-01-26 d931522
Use vapply rather than sapply
