### 2018-01-26 00c5daa
url = NULL uses default address
