### 2018-01-26 fc9e51c
Simplify preamble call
