+++ R/data.R
@@ -62,7 +62,7 @@
 #'   \item{wind_spd}{Wind speed km/hr}
 #'   \item{wind_spd_flag}{Wind speed flag}
 #'   \item{elev}{Elevation (m)}
-#'   \item{climat_id}{Climate identifier}
+#'   \item{climate_id}{Climate identifier}
 #'   \item{WMO_id}{World Meteorological Organization Identifier}
 #'   \item{TC_id}{Transport Canada Identifier}
 #' }
@@ -107,7 +107,7 @@
 #'   \item{wind_spd}{Wind speed km/hr}
 #'   \item{wind_spd_flag}{Wind speed flag}
 #'   \item{elev}{Elevation (m)}
-#'   \item{climat_id}{Climate identifier}
+#'   \item{climate_id}{Climate identifier}
 #'   \item{WMO_id}{World Meteorological Organization Identifier}
 #'   \item{TC_id}{Transport Canada Identifier}
 #' }
@@ -153,7 +153,7 @@
 #'   \item{total_snow}{Total snow (any form)}
 #'   \item{total_snow_flag}{Total snow flag}
 #'   \item{elev}{Elevation (m)}
-#'   \item{climat_id}{Climate identifier}
+#'   \item{climate_id}{Climate identifier}
 #'   \item{WMO_id}{World Meteorological Organization Identifier}
 #'   \item{TC_id}{Transport Canada Identifier}
 #' }
