+++ man/kamloops.Rd
@@ -37,7 +37,7 @@
   \item{wind_spd}{Wind speed km/hr}
   \item{wind_spd_flag}{Wind speed flag}
   \item{elev}{Elevation (m)}
-  \item{climat_id}{Climate identifier}
+  \item{climate_id}{Climate identifier}
   \item{WMO_id}{World Meteorological Organization Identifier}
   \item{TC_id}{Transport Canada Identifier}
 }}
