### 2018-01-29 ab548f2
Shorting line lengths and add global vars
