+++ R/weather.R
@@ -99,12 +99,18 @@ weather_dl <- function(station_ids,
                                  "climate_data/bulk_data_e.html")
 
   # Address as.POSIXct...
-  if((!is.null(start) & class(try(as.Date(start), silent = TRUE)) == "try-error") |
-     (!is.null(end) & class(try(as.Date(end), silent = TRUE)) == "try-error")) {
-    stop("'start' and 'end' must be either a standard date format (YYYY-MM-DD) or NULL")
+  if((!is.null(start) &
+      class(try(as.Date(start), silent = TRUE)) == "try-error") |
+     (!is.null(end) &
+      class(try(as.Date(end), silent = TRUE)) == "try-error")) {
+    stop("'start' and 'end' must be either a standard date format ",
+         "(YYYY-MM-DD) or NULL")
   }
 
-  if(length(interval) > 1) stop("'interval' must be either 'hour', 'day', OR 'month'")
+  if(length(interval) > 1) {
+    stop("'interval' must be either 'hour', 'day', OR 'month'")
+  }
+
   check_int(interval)
 
   tz_list <- c()
@@ -131,13 +137,15 @@ weather_dl <- function(station_ids,
 
     if(class(try(as.Date(stn1$start), silent = TRUE)) == "try-error") {
       stn1 <- dplyr::mutate(stn1,
-                            start = lubridate::floor_date(lubridate::ymd(as.character(start),
-                                                                         truncated = 2), "year"))
+                            start = lubridate::ymd(as.character(start),
+                                                   truncated = 2),
+                            start = lubridate::floor_date(start, "year"))
     }
     if(class(try(as.Date(stn1$end), silent = TRUE)) == "try-error") {
       stn1 <- dplyr::mutate(stn1,
-                            end = lubridate::ceiling_date(lubridate::ymd(as.character(end),
-                                                                         truncated = 2), "year"))
+                            end = lubridate::ymd(as.character(end),
+                                                 truncated = 2),
+                            end = lubridate::ceiling_date(end, "year"))
     }
     stn1 <- stn1 %>%
       dplyr::mutate(end = replace(end, end > Sys.Date(), Sys.Date()),
@@ -221,7 +229,9 @@ weather_dl <- function(station_ids,
                        as.character(lubridate::int_end(dates)), ").",
                        "\nAvailable Station Data:\n",
                        paste0(utils::capture.output(print(
-                         dplyr::filter(stn, station_id %in% station_ids, !is.na(start)))),
+                         dplyr::filter(stn,
+                                       station_id %in% station_ids,
+                                       !is.na(start)))),
                          collapse = "\n"))
     return(tibble::tibble())
   }
@@ -244,7 +254,9 @@ weather_dl <- function(station_ids,
                          as.character(lubridate::int_end(dates)), ").",
                          "\nAvailable Station Data:\n",
                          paste0(utils::capture.output(print(
-                           dplyr::filter(stn, station_id %in% station_ids, !is.na(start)))),
+                           dplyr::filter(stn,
+                                         station_id %in% station_ids,
+                                         !is.na(start)))),
                            collapse = "\n"))
       return(tibble::tibble())
     }
@@ -260,15 +272,18 @@ weather_dl <- function(station_ids,
 
     ## Appropriate grouping levels
     if(interval == "hour"){
-      w_all <- tidyr::nest(w_all, -station_name, -station_id, -lat, -lon, -date)
+      w_all <- tidyr::nest(w_all,
+                           -station_name, -station_id, -lat, -lon, -date)
     }
 
     if(interval == "day"){
-      w_all <- tidyr::nest(w_all, -station_name, -station_id, -lat, -lon, -month)
+      w_all <- tidyr::nest(w_all,
+                           -station_name, -station_id, -lat, -lon, -month)
     }
 
[…]
