### 2018-01-29 925e4e8
Add changes to NEWS
