### 2018-01-29 24a23d9
Use .data for multi-word objects rather than global variables
