### 2018-01-29 52d3f49
Match preamble cols even on windows with extra encoding characters
