### 2018-01-29 #160 @steffilazerte
Okay, I've removed the Date from the DESCRIPTION file, shortened all the too long lines and switched `vapply()` for `sapply()`. The master branch has all changes!
