### 2018-01-30 #160 @sckott
@steffilazerte closing issue now b/c approval, but can continue any discussion here still
