### 2018-01-31 #160 @steffilazerte
Okay sounds good, it asks me which teams should have access, I assume only two teams:  Leadership and weathercan?

- I'll update the links, add the footer and add metadata files after I transfer it (easiest way to make sure they work!)
- Definitely interested in blog post (my collaborator, @boshek, had an idea of showcasing weathercan in a spatial framework.)
- Where do we go with the JOSS manuscript?
