### 2018-01-31 #160 @sckott
> I assume only two teams

yep, thanks

> Definitely interested in blog post

Great, @stefaniebutland will get in touch with you

> Where do we go with the JOSS manuscript?

* Hook up your repo to Zenodo, see https://guides.github.com/activities/citable-code/ . Then generate a new release associated with a git tag in your repo. This doesn't mean you have to submit to CRAN with that git tag/release, its just to get a Zenodo DOI
* Then submit your `paper.md` to JOSS at http://joss.theoj.org/papers/new
