### 2018-01-31 #160 @stefaniebutland
@steffilazerte
> Definitely interested in blog post

Excellent! You can find some technical and editorial guidelines here: https://github.com/ropensci/roweb2#contributing-a-blog-post.

I have a spot open for a post on Tues Mar 6. We typically ask for a draft via pull request at least a week in advance so we can review it. What do you think about submitting a draft by Tues Feb 27th? If you submit your draft earlier, we might publish earlier if another post is delayed.

Thank you for being willing to do the extra work for a post. I hope it will get more eyes on your work.
