### 2018-02-01 #160 @steffilazerte
@sckott

- weathercan was transferred to ropensci
- I've updated all the links to ropensci (including the website and vignettes)
- Added ropensci footer
- added codemeta.json

But in trying to get weathercan set up with zenodo, I don't seem to have access to the weathercan repository in order to enable the hook. I also don't have access to the settings tab on weathercan anymore where the webhook details are stored. Is this something you need to enable?

Also can we get @boshek added to the weathercan team?
