### 2018-02-01 #160 @sckott
🚀 thanks

zenodo integration turned on and you and sam now have admin rights on repo
