### 2018-02-09 #160 @stefaniebutland
Hi @steffilazerte. Further thoughts on a date to submit a draft blog post? draft by Tues Feb 27th for publication the following week?
