### 2018-02-16 #160 @steffilazerte
Hi @stefaniebutland Okay, I think we've got it sorted :) Sam's got a lot going on, so I'll write the blog post solo.

As `weathercan` is pretty simple, I've been playing around with the idea of highlighting how to integrate data downloaded through weathercan into other datasets (as that's how most people would use it), be it spatial (e.g. with `sf` and then perhaps `mapview` or `leaflet`) or biological (e.g. interpolating the data to integrate it with bird activity data from my `feedr` package).

It's a pretty simple idea, but I find that it's usually the loading and combining data sets that most people get hung up on. Would this work? I'm not certain what style or length you're looking for.

I think I could get you a draft by March 9th, if that would work!
