### 2018-02-16 #160 @stefaniebutland
Your idea sounds good. You probably have the best handle on what users will want to do with the package and always good to write about something you're keen to share. Using the integration examples you suggest should allow others to figure out how to integrate with other data that interest them, so once you've written up your examples, see if there's a place for a paragraph pointing out any generic issues people should pay attention to in integrating with other datasets.

Maybe Sam can have a read-through before you submit pull request, as an informed second pair of eyes.

I have this post in my calendar to be published Tues Mar 13th, so could you submit draft by Mar 6th?
