### 2018-02-16 #160 @steffilazerte
Good idea about the generic issues, I'll definitely include that section. Glad you're up for the read through Sam! I should be able to have the draft by March 6th, no problem.
Thanks!
